+++ src/main/java/entitybuilder/GenerateBuyerUseCase.java
@@ -11,7 +11,7 @@ public class GenerateBuyerUseCase {
      * @param creditScore The car buyers credit score
      * @return
      */
-    public CarBuyer GenerateBuyerDataUseCase(double budget, int creditScore) {
+    public CarBuyer GenerateBuyerDataUseCase(int budget, int creditScore) {
         return new CarBuyer(budget, creditScore);
     }
 }
